Adjust layout to use dynamic pages
